Claim Detector Model/claim_detector/requirements_eve.txt
# Evidence & Value Engine (EVE) Dependencies

# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Storage
boto3>=1.34.0
supabase>=2.0.0
Pillow>=10.0.0

# OCR and PDF processing
pytesseract>=0.3.10
PyMuPDF>=1.23.0
opencv-python>=4.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dateutil>=2.8.2
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
